Fixed all known bugs.
Fixed a logic bug of slow motion; limited the interactive layer of left and right hand ray to avoid false teleportation.

It is known that there is a difference between the Android application file and the runtime numerical effect in the editor, and the multiplier of the throwing power is increased in the editor (normally it should be 5).